requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy==1.20.1
scipy==1.6.0
Pillow==8.1.0
scikit-image==0.18.1
scikit-learn==0.24.1
opencv-python==4.5.1.48
tensorboard==2.4.1
torchinfo==1.4.4
openslide-python==1.1.2
tqdm==4.61.2
